requirements.txt
python-miio
Flask
board
adafruit_dht